improved delay from 5 to 8 miliseconds,
Searching to quick into a new direction will result into ruined
discovery protocol, increasing this vallue to to high timout values will
slow down the discovery protocol. Because in the case that there are no
blocks into a direction we actualy have to wait 8 miliseconds. But we
dont want the case that we dont't wait to long for it to timeout.